Refactor mentions as a submodule of notifications
