+++ src/smc-webapp/notifications/index.ts
@@ -1,4 +1,4 @@
 export { NotificationPage } from "./notification-page";
 export { init } from "./init";
-export { MentionsActions } from "./actions";
-export { MentionsStore } from "./store";
+export { MentionsActions } from "./mentions/actions";
+export { MentionsStore } from "./mentions/store";
