+++ src/smc-webapp/notifications/init.ts
@@ -1,8 +1,8 @@
 import { AppRedux } from "../app-framework";
-import { MentionsStore } from "./store";
-import { MentionsActions } from "./actions";
-import { MentionsTable } from "./table";
-import { redux_name } from "./util";
+import { MentionsStore } from "./mentions/store";
+import { MentionsActions } from "./mentions/actions";
+import { MentionsTable } from "./mentions/table";
+import { redux_name } from "./mentions/util";
 
 export function init(redux: AppRedux) {
   if (redux.getStore(redux_name) != undefined) {
