+++ src/smc-webapp/notifications/mentions/index.ts
@@ -1,0 +1,4 @@
+export { MentionRow } from "./mention-row"
+export { MentionsActions } from "./actions";
+export { MentionsStore } from "./store";
+export { MentionTable } from "./table";
