+++ src/smc-webapp/notifications/mentions/util.ts
